+++ 10265/10-12_21-12-2012/src/Modulo14.Console/Program.cs
@@ -119,7 +119,7 @@
 				{
 					int id = int.Parse(dr["Id"].ToString());
 					string nome = dr["Nome"].ToString();
-					double preco = double.Parse(dr["Preco"].ToString());
+					decimal preco = decimal.Parse(dr["Preco"].ToString());
 					int quantidade = int.Parse(dr["Quantidade"].ToString());
 					bool ativo = bool.Parse(dr["Ativo"].ToString());
